Order Request Create Successfully

+++ src/main/java/com/food/zotatoFoods/entites/Order.java
@@ -1,6 +1,5 @@
 package com.food.zotatoFoods.entites;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -52,9 +51,9 @@ public class Order {
     private Consumer consumer;
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<OrderItem> orderItems = new ArrayList<>();
-    private BigDecimal foodAmount;
-    private BigDecimal platformFee;
-    private BigDecimal totalPrice;
+    private Double foodAmount;
+    private Double platformFee;
+    private Double totalPrice;
     @Column(columnDefinition = "Geometry(Point,4326)")
     private Point pickupLocation;
     @Column(columnDefinition = "Geometry(Point,4326)")
