+++ src/main/java/com/food/zotatoFoods/entites/WalletTransaction.java
@@ -1,6 +1,5 @@
 package com.food.zotatoFoods.entites;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.*;
@@ -26,7 +25,7 @@ public class WalletTransaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private BigDecimal Amount;
+    private Double Amount;
     @Enumerated(EnumType.STRING)
     private TransactionType transactionType;
     @Enumerated(EnumType.STRING)
